requirements.txt
google-api-python-client
google-auth
py-dmidecode
python-dotenv